% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_iso_month}
\alias{get_iso_month}
\title{Funcion para obtener el mes para un determinado isoweek}
\usage{
get_iso_month(years, weeks)
}
\arguments{
\item{years}{año actual}

\item{weeks}{semana iso a consultar}
}
\value{
isoweek
}
\description{
Funcion para obtener el mes para un determinado isoweek
}
